+++ lib/src/array2d/numeric_axis.dart
@@ -13,23 +13,28 @@ abstract class Numeric2DAxisFix<E extends num>
 abstract class Numeric2DAxisView<E extends num> implements Axis2DView<E> {
   Numeric1DView<E> operator [](int r);
 
-  Array<double> get mean;
+  Numeric1D<double> get mean;
 
-  Array<E> get sum;
+  Numeric1D<E> get sum;
 
-  Array<double> average(Iterable<num> weights);
+  Numeric1D<double> average(Iterable<num> weights);
 
-  Array<double> get variance;
+  Numeric1D<double> get variance;
 
-  Array<double> get std;
+  Numeric1D<double> get std;
 
-  Numeric2D<E> operator +(Numeric1DView<E> other);
+  Numeric2D<E> operator +(
+      /* num | IterView<num> | Iterable<num> | Numeric2D<int> */ other);
 
-  Numeric2D<E> operator -(Numeric1DView<E> other);
+  Numeric2D<E> operator -(
+      /* num | IterView<num> | Iterable<num> | Numeric2D<int> */ other);
 
-  Numeric2D<E> operator *(Numeric1DView<E> other);
+  Numeric2D<E> operator *(
+      /* num | IterView<num> | Iterable<num> | Numeric2D<int> */ other);
 
-  Numeric2D<double> operator /(Numeric1DView other);
+  Numeric2D<double> operator /(
+      /* num | IterView<num> | Iterable<num> | Numeric2D<int> */ other);
 
-  Numeric2D<int> operator ~/(Numeric1DView other);
+  Numeric2D<int> operator ~/(
+      /* num | IterView<num> | Iterable<num> | Numeric2D<int> */ other);
 }