+++ lib/src/series/bool.dart
@@ -1,3 +1,9 @@
 part of grizzly.primitives.series;
 
-abstract class BoolSeriesBase<LT> implements Series<LT, bool> {}
+abstract class BoolSeriesBase<LT>
+    implements Series<LT, bool>, BoolSeriesFixBase<LT> {}
+
+abstract class BoolSeriesFixBase<LT>
+    implements SeriesFix<LT, bool>, BoolSeriesViewBase<LT> {}
+
+abstract class BoolSeriesViewBase<LT> implements SeriesView<LT, bool> {}