+++ lib/src/series/series.dart
@@ -17,7 +17,15 @@ typedef Series<LT, VT> SeriesMaker<LT, VT>(Iterable<VT> data,
 
 /// A series with index of type [LT] and value of type [VT]
 abstract class Series<LT, VT> implements SeriesFix<LT, VT> {
+  /// Name of the series
+  // TODO String name;
+
+  /// Data of the series
+  ArrayFix<VT> get data;
+
   void append(LT label, VT value);
+
+  void insert(int pos, LT label, VT value);
 
   /// Remove element at position [position]
   void remove(int position);
@@ -29,7 +37,7 @@ abstract class Series<LT, VT> implements SeriesFix<LT, VT> {
   void drop(LT label);
 
   /// Drop elements by label [label]
-  void dropMany(List<LT> label);
+  void dropMany(/* Iterable<LT> | IterView<LT> | Labeled */ label);
 
   void assign(/* Series<LT, VT> | IterView<VT> */ other, {bool addNew: true});
 
@@ -39,11 +47,36 @@ abstract class Series<LT, VT> implements SeriesFix<LT, VT> {
 
   void sortByLabel({bool descending: false});
 
-  void mask(IterView<bool> mask);
+  void keep(mask);
+
+  void keepOnly(Labeled<LT> mask);
+
+  void keepLabels(/* Iterable<LT> | IterView<LT> */ mask);
+
+  void keepIf(BoolSeriesViewBase<LT> mask);
+
+  void keepWhen(SeriesCond<LT, VT> cond);
+
+  // TODO addColumn
+
+  // TODO removeColumn
+
+  NumericSeries<LT, int> get asInt;
+
+  NumericSeries<LT, double> get asDouble;
+
+  BoolSeriesBase<LT> get asBool;
+
+  StringSeriesBase<LT> get asString;
+
+  DynamicSeriesBase<LT> get asDynamic;
 }
 
 /// A series with index of type [LT] and value of type [VT]
 abstract class SeriesFix<LT, VT> implements SeriesView<LT, VT> {
+  /// Data of the series
+  ArrayFix<VT> get data;
+
   /// Sets [value] by [label]
   operator []=(LT label, VT value);
 
@@ -60,4 +93,16 @@ abstract class SeriesFix<LT, VT> implements SeriesView<LT, VT> {
   void assignMap(Map<LT, VT> other);
 
   SeriesFix<LT, VT> get fixed;
+
+  NumericSeriesFix<LT, int> get asInt;
+
+  NumericSeriesFix<LT, double> get asDouble;
+
+  BoolSeriesFixBase<LT> get asBool;
+
+  StringSeriesFixBase<LT> get asString;
+
+  DynamicSeriesFixBase<LT> get asDynamic;
 }
+
+typedef bool SeriesCond<LT, VT>(LT lab, SeriesView<LT, VT> s);