+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.2.3
+
++ Has breaking changes
++ Better arithmetic operator methods for Array2D
+
+## 0.2.4
+
++ Several improvements to Array and Array2D
+
+## 0.2.1
+
++ New API
+
 ## 0.1.0
 
 + New API